# docs/openapi.yaml
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Bucket API","description":"API to manage files stored in a bucket (upload, download, update, delete, list, share)"}],"paths":{"/api/objects":{"get":{"tags":["Bucket API"],"summary":"List files in a directory","description":"Returns a list of files located in a directory within the bucket","operationId":"list","parameters":[{"name":"path","in":"query","description":"Directory path in the bucket","required":true,"schema":{"type":"string"},"example":"/path/in/bucket"}],"responses":{"200":{"description":"List of files returned successfully","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}},"500":{"description":"Internal server error"}}},"put":{"tags":["Bucket API"],"summary":"Update a file","description":"Replaces an existing file in the bucket with a new local file","operationId":"update","parameters":[{"name":"localPath","in":"query","description":"Path to the local file","required":true,"schema":{"type":"string"},"example":"/path/to/local/file.txt"},{"name":"remoteSrc","in":"query","description":"Path of the file to update in the bucket","required":true,"schema":{"type":"string"},"example":"/path/in/bucket/file.txt"}],"responses":{"204":{"description":"File updated successfully"},"404":{"description":"File not found"},"500":{"description":"Internal server error"}}},"post":{"tags":["Bucket API"],"summary":"Upload a file","description":"Uploads a local file to the bucket at the specified remote path","operationId":"upload","parameters":[{"name":"localPath","in":"query","description":"Path to the local file to upload","required":true,"schema":{"type":"string"},"example":"/path/to/local/file.txt"},{"name":"remoteSrc","in":"query","description":"Destination path in the bucket","required":true,"schema":{"type":"string"},"example":"/path/in/bucket/file.txt"}],"responses":{"201":{"description":"File uploaded successfully"},"400":{"description":"Invalid request parameters"},"500":{"description":"Internal server error"}}},"delete":{"tags":["Bucket API"],"summary":"Delete a file or directory","description":"Deletes a file or directory from the bucket, optionally recursively","operationId":"delete","parameters":[{"name":"remoteSrc","in":"query","description":"Path of the file or directory in the bucket","required":true,"schema":{"type":"string"},"example":"/path/in/bucket/file.txt"},{"name":"recursive","in":"query","description":"Whether the deletion should be recursive (for directories)","required":false,"schema":{"type":"boolean","default":false},"example":false}],"responses":{"204":{"description":"Deletion completed successfully"},"404":{"description":"File not found"},"500":{"description":"Internal server error"}}}},"/api/objects/share":{"post":{"tags":["Bucket API"],"summary":"Share a file","description":"Generates a temporary public URL to access a file stored in the bucket","operationId":"share","parameters":[{"name":"remoteSrc","in":"query","description":"Path of the file in the bucket","required":true,"schema":{"type":"string"},"example":"/path/in/bucket/file.txt"},{"name":"expirationTime","in":"query","description":"Expiration time in seconds","required":true,"schema":{"type":"integer","format":"int32"},"example":3600}],"responses":{"200":{"description":"Shared URL generated successfully","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"File not found"},"500":{"description":"Internal server error"}}}},"/api/objects/download":{"get":{"tags":["Bucket API"],"summary":"Download a file","description":"Downloads a file from the bucket to a local destination","operationId":"download","parameters":[{"name":"localPath","in":"query","description":"Local path where the file will be saved","required":true,"schema":{"type":"string"},"example":"/path/to/local/file.txt"},{"name":"remoteSrc","in":"query","description":"Path of the file in the bucket","required":true,"schema":{"type":"string"},"example":"/path/in/bucket/file.txt"}],"responses":{"200":{"description":"File downloaded successfully"},"404":{"description":"File not found"},"500":{"description":"Internal server error"}}}}},"components":{}}